+++ src/types.rs
@@ -1,11 +1,11 @@
 //! MQTT-specific data types used by the public API.
 use crate::QoS;
-use bit_field::BitField;
 use serde::Serialize;
 use serde::ser::SerializeStruct;
 
 /// MQTT binary data field.
-#[derive(defmt::Format, Copy, Clone, Debug, PartialEq)]
+#[derive(Copy, Clone, Debug, PartialEq)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub(crate) struct BinaryData<'a>(pub(crate) &'a [u8]);
 
 impl serde::Serialize for BinaryData<'_> {
@@ -103,7 +103,8 @@ impl<'a> Auth<'a> {
 }
 
 /// MQTT UTF-8 string field.
-#[derive(defmt::Format, Copy, Clone, Debug, PartialEq)]
+#[derive(Copy, Clone, Debug, PartialEq)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub(crate) struct Utf8String<'a>(pub(crate) &'a str);
 
 impl serde::Serialize for Utf8String<'_> {
@@ -203,11 +204,14 @@ impl SubscriptionOptions {
 
 impl serde::Serialize for SubscriptionOptions {
     fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        let value = *0u8
-            .set_bits(0..2, self.maximum_qos as u8)
-            .set_bit(2, self.no_local)
-            .set_bit(3, self.retain_as_published)
-            .set_bits(4..6, self.retain_behavior as u8);
+        let mut value = (self.maximum_qos as u8) & 0b11;
+        if self.no_local {
+            value |= 1 << 2;
+        }
+        if self.retain_as_published {
+            value |= 1 << 3;
+        }
+        value |= (self.retain_behavior as u8) << 4;
         serializer.serialize_u8(value)
     }
 }